refactor(application): 패키지 정리
작업내용
- 잘못된 위치에 배치된 패키지를 정상 위치에 배치
