+++ src/main/java/attendance/domain/AttendanceSystem.java
@@ -3,11 +3,11 @@ package attendance.domain;
 import attendance.domain.checker.AttendanceChecker;
 import attendance.domain.checker.AttendanceType;
 import attendance.domain.crew.CrewStorage;
-import attendance.domain.dto.AttendanceState;
-import attendance.domain.dto.RecordUpdateResult;
+import attendance.domain.record.AttendanceRecord;
 import attendance.domain.risk.RiskType;
+import attendance.dto.AttendanceState;
+import attendance.dto.RecordUpdateResult;
 import attendance.exception.ExceptionMessage;
-import attendance.record.AttendanceRecord;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
