+++ src/test/java/attendance/AttendanceSystemTest.java
@@ -9,10 +9,10 @@ import attendance.domain.checker.AttendanceChecker;
 import attendance.domain.checker.AttendanceType;
 import attendance.domain.checker.HolidayChecker;
 import attendance.domain.crew.CrewStorage;
-import attendance.domain.dto.AttendanceState;
+import attendance.domain.record.AttendanceRecord;
 import attendance.domain.risk.RiskType;
+import attendance.dto.AttendanceState;
 import attendance.exception.ExceptionMessage;
-import attendance.record.AttendanceRecord;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
